requirements.txt
Django==4.2.7
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
zope.interface==6.1
channels-redis==4.1.0
selenium==4.9.1
pytest-django==4.6.0
pytest-splinter==3.3.2
pytest-factoryboy==2.6.0
django-environ==0.11.2
pytest-mock==3.12.0
libsass==0.22.0
django_compressor==4.4
django-libsass==0.9